+++ NewSerial/src/main/java/com/example/newserial/domain/member/repository/Member.java
@@ -1,7 +1,7 @@
 package com.example.newserial.domain.member.repository;
 
 import com.example.newserial.domain.bookmark.repository.Bookmark;
-import com.example.newserial.domain.memo.entity.Memo;
+import com.example.newserial.domain.memo.repository.Memo;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
