+++ NewSerial/src/main/java/com/example/newserial/domain/memo/controller/MemoController.java
@@ -1,8 +1,9 @@
-package com.example.newserial.domain.memo;
+package com.example.newserial.domain.memo.controller;
 
 import com.example.newserial.domain.memo.dto.MemoResponseDto;
 import com.example.newserial.domain.memo.dto.MemoSaveRequestDto;
 import com.example.newserial.domain.memo.dto.MemoUpdateRequestDto;
+import com.example.newserial.domain.memo.service.MemoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
