+++ NewSerial/src/main/java/com/example/newserial/domain/memo/repository/MemoRepository.java
@@ -1,7 +1,5 @@
-package com.example.newserial.domain.memo;
+package com.example.newserial.domain.memo.repository;
 
-import com.example.newserial.domain.memo.repository.Memo;
-import com.example.newserial.domain.memo.repository.MemoId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemoRepository extends JpaRepository<Memo, MemoId> {
